pyproject.toml
[project]
name = "crehana-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "fastapi[standard]>=0.116.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.10.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.24",
    "strawberry-graphql[fastapi]>=0.278.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "coverage>=7.10.2",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["F", "N", "W", "I001", "I002"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-v --tb=short --disable-warnings"
testpaths = ["tests"]
python_files = "test_*.py"
pythonpath = "."

[tool.coverage.run]
branch = "true"  
source = ["app"]

[tool.coverage.report]
show_missing = "true"  
omit = ["tests/*", "*/__init__.py"]
fail_under = 85
